Add support for env var
