--- conftest.py
@@ -1,24 +1,0 @@
-# This code is part of OpenFE and is licensed under the MIT license.
-# For details, see https://github.com/OpenFreeEnergy/openfe
-import pytest
-
-# allow for optional slow tests
-# See: https://docs.pytest.org/en/latest/example/simple.html
-def pytest_addoption(parser):
-    parser.addoption(
-        "--runslow", action="store_true", default=False, help="run slow tests"
-    )
-
-
-def pytest_configure(config):
-    config.addinivalue_line("markers", "slow: mark test as slow to run")
-
-
-def pytest_collection_modifyitems(config, items):
-    if config.getoption("--runslow"):
-        # --runslow given in cli: do not skip slow tests
-        return
-    skip_slow = pytest.mark.skip(reason="need --runslow to run")
-    for item in items:
-        if "slow" in item.keywords:
-            item.add_marker(skip_slow)
